+++ scripts/build_retro_cost.py
@@ -28,6 +28,9 @@ The energy savings calculations are based on the
       - tabula https://episcope.eu/fileadmin/tabula/public/calc/tabula-calculator.xlsx
 
 
+The energy savings for hot water are based on this report by the University of Innsbruck:
+https://diglib.uibk.ac.at/ulbtirolfodok/content/titleinfo/7640369
+
 Basic Equations
 ---------------
 
@@ -724,8 +727,10 @@ def map_to_lstrength(l_strength, df):
     )
     # reflects in capital costs, otherwise moderate cost assumptions are too high,
     # compared to ambitious ones (this setting lowers them by approx. 10%)
+    # previously: middle = len(l_strength) // 2
     map_to_l = pd.MultiIndex.from_arrays(
         [middle * [3] + len(l_strength[middle:]) * [3], l_strength]
+        # previously: [middle * [2] + len(l_strength[middle:]) * [3], l_strength]
     )
     l_strength_df = (
         df.stack(-2)
@@ -1047,6 +1052,37 @@ def sample_dE_costs_area(
     return cost_dE_new, area_tot
 
 
+def WWHR_costs(households):
+    """
+    Calculates the costs for waste water heat recovery (WWHR) based on the
+    number of households per region.
+    """
+    pop_layout = pd.read_csv(snakemake.input.clustered_pop_layout, index_col=0)
+    housholds_spatial = pd.merge(
+        pop_layout.reset_index(), households, on="ct"
+    ).set_index("name")
+    housholds_spatial = (
+        housholds_spatial.fraction * housholds_spatial["Households (thousands)"] * 1000
+    )  # number in thousands
+    costs_WWHR = (
+        pd.read_csv(snakemake.input.cost_germany, index_col=0, usecols=[0, 1, 2, 3])
+        .loc["hot water (WWHRS)"]
+        .astype(float)
+    )
+
+    if annualise_cost:
+        if interest_rate > 0:
+            costs_WWHR = interest_rate / (
+                1.0 - 1.0 / (1.0 + interest_rate) ** costs_WWHR.loc["life_time"]
+            )
+        else:
+            costs_WWHR = 1 / costs_WWHR.loc["life_time"]
+
+    costs_WWHR = costs_WWHR * housholds_spatial
+
+    return costs_WWHR
+
+
 # %% --- MAIN --------------------------------------------------------------
 if __name__ == "__main__":
     if "snakemake" not in globals():
@@ -1062,6 +1098,10 @@ if __name__ == "__main__":
     set_scenario_config(snakemake)
 
     #  ********  config  *********************************************************
+
+    households = pd.read_csv(snakemake.input.households).rename(
+        columns={"Country": "ct"}
+    )
 
     retro_opts = snakemake.params.retrofitting
     interest_rate = retro_opts["interest_rate"]
@@ -1109,6 +1149,10 @@ if __name__ == "__main__":
         area, area_tot, costs, dE_space, countries, construction_index, tax_weighting
     )
 
+    # Calculare the costs for waste water heat recovery
+    WWHR_costs = WWHR_costs(households)
+
     #   save *********************************************************************
     cost_dE.to_csv(snakemake.output.retro_cost)
     area_tot.to_csv(snakemake.output.floor_area)
+    WWHR_costs.to_csv(snakemake.output.WWHR_costs)